Add test for isEmpty
Bags that have things added to them, such as those in our test, should not be empty. The display name can be customized to show what you want for the label for the test. The `@Test` annotation is telling us that when we run tests, this is one of them. the test as one assertion - it asserts that the result of running the method isEmpty() for bigBag2 should be false.