Add test for removing a specific item
Testing this method made me realize I wanted to have a BeforeEach as well as a BeforeAll.
I moved some things from the BeforeAll to a BeforeEach, and I made things not static that
were in the BeforeEach (along with that method). I'm not sure of best practices here,
but it struck me that having the bag be static might not make much sense, while items being static seemed fine.